requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
spacy==3.7.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
jinja2==3.1.4
redis==5.0.8
python-dotenv==1.0.1
openai==1.52.2

